project/gradio/requirements.txt
gradio==5.23
requests
pillow